Remove RETAIN() when returning -[NSPasteboardItem types].
Objective-C memory management policy specifies that methods named with
alloc, new, copy and mutableCopy are the ones that return objects with
refcount 1.

Additionally, `@property (..., copy, ...)` only modifies the setter; the getter
still returns the object as-is. While it should be correct to create
a new immutable object every time, we should not create an object
with refcount 1, but an autoreleased object.