NSBox: (re-)implement isOpaque
This (re-)implements logic where a box is marked
as opaque if it is a custom box which is not
transparent, but moves the logic to `GSTheme`,
allowing themes to override this loggic.

This fixes the rendering of custom boxes with a
custom fill color; the current implementation
would mark them as non-opaque and hence the fill
color would not be rendered.

- The original implementation (from 1999) always
  returned `YES`
  (53bcb5024020fe85e370d9d49297aca4320d7768)
- A more complete implementation was backported
  from the Testplant branch in 2012
  (e85b16bc05d5d513788c09b470fd09ad5f888882, by
  @fredkiefer, upstream commit
  36e77b95f7921e9539fdc0c115fe4039fa573150)
- More recently, the a return value of `NO` was
  hardcoded
  (02bc49e2d5336e9092001e3d56cb43750a8466a3, by
  @ericwa)