Add APPKIT_EXPORT_CLASS, APPKIT_EXPORT annotations (#288)
Some GS* classes and methods are used by the ColorPicker or Printing bundles.  That means that they are referenced from an external module.  For this module to be able to resolve these symbols, on Windows, they need to be exported.

This commit adds `APPKIT_EXPORT_CLASS` and `APPKIT_EXPORT` annotations, allowing ColorPickers to compile on Windows.