always reset _needsLayout in -layout and not only -layoutSubtreeIfNeeded via layoutViewAndSubvies.
Apple states for -(void) layout:

Override this method if your custom view needs to perform custom layout not expressible using the constraint-based layout system. In this case you are responsible for setting needsLayout to true when something that impacts your custom layout changes.

Which seems to suggest that layout does the reset of needsLayout, since
setting it is up to programmer.